w3c-reports/12092014/gtk/runned-tests/html/webappapis/scripting/events/event-handler-spec-example-expected.txt


FAIL Event handler listeners should be registered when they are first set to an object value (object "[object Object]"). assert_equals: expected 2 but got 1(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:18:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:23:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:13:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:9:16)
FAIL Event handler listeners should be registered when they are first set to an object value (function "function () {}"). assert_equals: expected 2 but got 1(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:18:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:23:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:13:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:9:16)
FAIL Event handler listeners should be registered when they are first set to an object value (object "42"). assert_equals: expected 2 but got 1(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:18:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:23:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:13:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:9:16)
FAIL Event handler listeners should be registered when they are first set to an object value (object ""). assert_equals: expected 2 but got 1(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:18:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:23:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:13:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:9:16)
FAIL Event handler listeners should be registered when they are first set to an object value (42). assert_equals: expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:40:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:43:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:33:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:29:19)
FAIL Event handler listeners should be registered when they are first set to an object value (null). assert_equals: expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:40:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:43:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:33:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:29:19)
FAIL Event handler listeners should be registered when they are first set to an object value (undefined). assert_equals: expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:40:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:43:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:33:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:29:19)
FAIL Event handler listeners should be registered when they are first set to an object value (""). assert_equals: expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:40:77
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:43:17
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:33:9
forEach@[native code]
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:29:19)
FAIL Event handler listeners should be registered when they are first set to an object value. assert_equals: expected 3 but got 2(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:57:75
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
click@[native code]
http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:60:15
step@http://localhost:8800/resources/testharness.js:798:30
global code@http://localhost:8800/html/webappapis/scripting/events/event-handler-spec-example.html:51:7)
